requirements.txt
Django>=4.0,<6.0
Pillow>=9.0.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
python-decouple>=3.8
psycopg2-binary>=2.9
gunicorn>=21.0
whitenoise>=6.4
django-filter>=23.0
pandas>=2.0
numpy>=1.24
openpyxl>=3.1
xlrd>=2.0
dj-database-url>=2.0.0
python-dotenv>=1.0.0
supabase>=2.0.0
django-model-utils>=4.3.1
django-widget-tweaks>=1.5.0
django-taggit>=4.0.0
pdfkit>=0.6.0 